requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
blinker==1.6.2
    # via flask
click==8.1.3
    # via flask
decorator==5.1.1
    # via validators
flask==2.3.2
    # via
    #   -r requirements.in
    #   flask-basicauth
    #   flask-sqlalchemy
flask-basicauth==0.2.0
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.0.0
    # via -r requirements.in
short-url==1.2.2
    # via -r requirements.in
sqlalchemy==2.0.15
    # via flask-sqlalchemy
typing-extensions==4.5.0
    # via sqlalchemy
uwsgi==2.0.21
    # via -r requirements.in
validators==0.20.0
    # via -r requirements.in
werkzeug==2.3.4
    # via flask
